Move core functionality into separate hook
